requirements.txt
# Data handling
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
fastparquet>=2023.0

# Configuration
pyyaml>=6.0

# Machine learning
scikit-learn>=1.3

# Gradient boosting
xgboost>=2.0
lightgbm>=4.0

# Deep learning (ANN)
torch>=2.0

# Hyperparameter tuning
optuna>=3.0

# Interpretability
shap>=0.43

# Visualization
matplotlib>=3.7
seaborn>=0.13

# Notebook support
jupyter>=1.0
ipykernel>=6.0

# Testing
pytest>=7.0

# Statistics
scipy>=1.11
statsmodels>=0.14
